Make plugins objects, not classes
Before, all plugins were passed as a class, and all of their actions were class methods. Now you pass a plugin as an instance of the plugin class.

Here’s why: there’s no reason why there couldn’t be multiple Goldengates in an app. And in that case, you’d probably want the plugins to be parametrized and pass some stuff to the plugin initializer.